Board 클래스를 분리
- Board : 보드 위 기물들의 위치를 저장하고 , 상태를 표현할 수 있다
- ChessGame : 체스 게임 진행을 위한 동작을 실행할 수 있다 (move 등)
- ChessHelp : 체스 게임 룰 , 위치 문자열을 배열의 인덱스로 변환하는 메서드 제공
